Use native stored ZIPs for support bundles
